% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting.R
\name{balancedTrainProb}
\alias{balancedTrainProb}
\title{Balanced probablity that an observation should be assigned to a certain set}
\usage{
balancedTrainProb(i, split, X)
}
\arguments{
\item{i}{is an index of the added object in X}

\item{split}{is the indices of observations from Train and Holdout sets}

\item{X}{the covariates}
}
\value{
A balansed probablity that the observation i should be in the train dataset
}
\description{
Balanced probablity that an observation should be assigned to a certain set
}
