% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting.R
\name{balansedSplit}
\alias{balansedSplit}
\title{Balansed proportional train-holdout splitting of a dataset}
\usage{
balansedSplit(opts, dbY, dbTrt, dbX)
}
\arguments{
\item{dbY}{is Y componnet of the dataset}

\item{dbTrt}{is Trt componnet of the dataset}

\item{dbX}{is X componnet of the dataset}

\item{opts$InitDataPropotion}{defines proportion of a dataset that is randomly assigned to train and holdout sets}

\item{opts$InitSplitPropotion}{defines proportion of initial split into train and holdout sets}

\item{opts$Deterministic}{is a logical flag defining if balancing assigniments should be deterministic}
}
\value{
A list with indices of train and holdout set.
}
\description{
The function realizes the procedure described in
Lipkovich I. et al. Subgroup identification based on differential effect search -- A recursive partitioning method for establishing response to treatment in patient subpopulations // Stat. Med. John Wiley & Sons, Ltd, 2011. Vol. 30, № 21. P. 2601–2621.
}
\examples{

# Generating dataset
Trt = rbinom(1000,1,0.5)
X = data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000))
Y = ( 2*X$X1 - 1 + X$X2*Trt + rnorm(1000) ) > 0
# Computing the train-holdout split
split = balansedSplit(list(InitSplitProportion=0.2), Y, Trt, X)
summary(split)
summary(X[split$Train,])
summary(X[split$Holdout,])
}
