% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidateAlgos_par.R
\name{crossValidateAlgos_par}
\alias{crossValidateAlgos_par}
\title{A CrossValidateAlgos_par Function}
\usage{
crossValidateAlgos_par(
  models,
  subgroupQualityFuncs,
  quantile.probs,
  dbY,
  dbTrt,
  dbX,
  numTrials,
  splitFunc,
  splitOpts
)
}
\arguments{
\item{models}{is a vector of model descriptios that contains Train function, Train options, Predict function and the name of the model.
The prototype of the train function is function(Y,Trt,X, opts),
  where Y is the response variable, Trt is the 0-1 treatment variable, X is the covariate matrix, and opts is the options from the corresponding model.
The prototype for the Predict function is function(m,X) where m is a trained model and X the observations for prediction.}

\item{subgroupQualityFuncs}{is a vector of functions that evaluate the quality of a subgroup. The prototype is function(subgroup, Y, Trt), where subgroup is T-F vector defining a subgroup, and Y and Trt are similar as for the functions from trainModelFuncs}

\item{quantile.probs}{is a vector of probabilities for quantile in order to determine subgroups where an effect should be computed}

\item{trainY}{a train response variable}

\item{trainTrt}{a train treatment 0-1 variable}

\item{trainX}{train covariates}

\item{holdoutY}{a holdout response variable}

\item{holdoutTrt}{a holdout treatment 0-1 variable}

\item{holdoutX}{holdout covariates}
}
\value{
a list with found subgroups in the holdout set, their sizes, qualities, and qualities of a random subset of similar size for all the algorithms
}
\description{
This function allows you to parallel evaluation.
}
\examples{
crossValidateAlgos_par()

}
