% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-wrap-up-functions.R
\name{modifyDataByTian}
\alias{modifyDataByTian}
\title{Covariate transformation as defined by Tian et al., 2014}
\usage{
modifyDataByTian(Trt, X)
}
\arguments{
\item{Trt}{the binary treatment varibale}

\item{X}{the covariates}
}
\value{
the modified covariate matrix
}
\description{
Covariate transformation as defined by Tian et al., 2014
}
