% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-wrap-up-functions.R
\name{modifyDataByWeisberg}
\alias{modifyDataByWeisberg}
\title{Outcome transformation as defined by Weisberg & Pontes, 2015}
\usage{
modifyDataByWeisberg(Trt, Y)
}
\arguments{
\item{Trt}{is the binary treatment variable}

\item{Y}{is the binary response varibale}
}
\value{
the modified outcome matrix
}
\description{
Outcome transformation as defined by Weisberg & Pontes, 2015
}
