% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict2MXGb}
\alias{predict2MXGb}
\title{Predict function for train2MXGb}
\usage{
predict2MXGb(m, X)
}
\arguments{
\item{m}{a model trained with train2MXGb}

\item{X}{is the covariate matrix}
}
\value{
a vector with predicted PTE
}
\description{
Predicts PTE of observations by means difference in predictions between two independent XGboost models (Predict function for train2MXGb)
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fitting model
m = train2MXGb(Y, Trt, X, opts = list(Trt = list(nrounds = 15, eta = 0.3, subsample = 0.5, depth = 4), Non_Trt = list(nrounds = 10, eta = 0.4, subsample = 0.75, depth = 3)))
# Predicting PTE
predict2MXGb(m, data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000)))

}
