% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-wrap-up-functions.R
\name{predictByModelModLM}
\alias{predictByModelModLM}
\title{Predicts if an observation should be treated or not by means of the model trained with trainModelModLM}
\usage{
predictByModelModLM(m, X)
}
\arguments{
\item{m}{a model trained with trainModelModLM}

\item{X}{is the covariate matrix}
}
\value{
a binary vector specifying the observations to be treated
}
\description{
Predicts if an observation should be treated or not by means of the model trained with trainModelModLM
}
\examples{

# Generating dataset
Trt = rbinom(1000,1,0.5)
X = data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(1000) ) > 0 )
# Training
m = trainModelModLM(Y,Trt,X)
# Predicting a good group for new data
sum(predictByModelModLM(m, data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000))))

}
