% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictTianGLM}
\alias{predictTianGLM}
\title{Predict function for trainTianGLM}
\usage{
predictTianGLM(m, X)
}
\arguments{
\item{m}{a model trained with trainTianGLM}

\item{X}{is the covariate matrix}
}
\value{
a vector with predicted PTE
}
\description{
Predicts PTE of observations by means of the generalized linear model where covariate transformation as defined by Tian et al., 2014 is used (Predict function for trainTianGLM)
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fitting model
m = trainTianGLM(Y, Trt, X, opts = list(alpha = 0, lambda = 0.15))
# Predicting PTE
predictTianGLM(m, data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000)))

}
