% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictUpliftModelRF}
\alias{predictUpliftModelRF}
\title{Predict function for trainUpliftModelRF}
\usage{
predictUpliftModelRF(m, X)
}
\arguments{
\item{m}{a model trained with trainUpliftModelRF}

\item{X}{is the covariate matrix}
}
\value{
a vector with predicted PTE
}
\description{
Predicts PTE of observations by means difference in predictions between two independent Uplift Random forest (Guelman, 2014) models (Predict function for trainUpliftModelRF)
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fitting model
m = trainUpliftModelRF(Y, Trt, X, opts = list(split_method = "Chisq", ntree = 55, bag.fraction = 0.5))
# Predicting PTE
predictUpliftModelRF(m, data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000)))

}
