% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictWeisbergGLM}
\alias{predictWeisbergGLM}
\title{Predict function for trainWeisbergGLM}
\usage{
predictWeisbergGLM(m, X)
}
\arguments{
\item{m}{a model trained with trainWeisbergGLM}

\item{X}{is the covariate matrix}
}
\value{
a vector with predicted PTE
}
\description{
Predicts PTE of observations by means of the generalized linear model where binary response transformation as defined by Weisberg et al., 2015 is used (Predict function for trainWeisbergGLM)
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fitting model
m = trainWeisbergGLM(Y, Trt, X, opts = list(alpha = 0, lambda = 0.15))
# Predicting PTE
predictWeisbergGLM(m, data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000)))

}
