% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictWeisbergRF}
\alias{predictWeisbergRF}
\title{Predict function for trainWeisbergRF}
\usage{
predictWeisbergRF(m, X)
}
\arguments{
\item{m}{a model trained with trainWeisbergRF}

\item{X}{is the covariate matrix}
}
\value{
a vector with predicted PTE
}
\description{
Predicts PTE of observations by means of the Random Forest model where binary response transformation as defined by Weisberg et al., 2015 is used (Predict function for trainWeisbergRF)
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fitting model
m = trainWeisbergRF(Y, Trt, X, opts = list(mtry = 2, ntree = 15, nodesize = 200))
# Predicting PTE
predictWeisbergRF(m, data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000)))

}
