% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting.R
\name{randomSplit_equal}
\alias{randomSplit_equal}
\title{Random balanced proportional train-holdout splitting of a dataset}
\usage{
randomSplit_equal(opts, dbY, dbTrt, dbX)
}
\arguments{
\item{dbY}{is Y component of the dataset}

\item{dbTrt}{is Trt component of the dataset}

\item{dbX}{is X component of the dataset}

\item{opts$TestPropotion}{defines the relative size of the test set}
}
\value{
A list with indices of train and holdout set.
}
\description{
Takes a dataset and randomly assign observations to either train or holdout sets in a given proportion and providing equal shares of treated (Trt = 1) and control observations (Trt = 0) both in train and holdout sets
}
\examples{
# Generating dataset
Trt = rbinom(1000,1,0.5)
X = data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000))
Y = ( 2*X$X1 - 1 + X$X2*Trt + rnorm(1000) ) > 0
# Computing the train-holdout split
split = randomSplit_equal(list(TestProportion=0.2), Y, Trt, X)
}
