% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudite.R
\docType{package}
\name{SuDiTE}
\alias{SuDiTE}
\alias{SuDiTE-package}
\title{SuDiTE: A package for comaparison of different uplift models}
\description{
It includes comparison infrastructure as well as wrap-ups for many existing algos
}
\details{
It is written as a part of SuDiTE project in the National Research University Higher School of Economics, Perm, Russia
}
