% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-wrap-up-functions.R
\name{trainModelModLM}
\alias{trainModelModLM}
\title{Model by Tian et al 2014}
\usage{
trainModelModLM(Y, Trt, X)
}
\arguments{
\item{Y}{is the binary response varibale}

\item{Trt}{is the binary treatment variable}

\item{X}{is the numeric covariates matrix}
}
\value{
a model that is able to classify clients into treatment and control group
}
\description{
Model by Tian et al 2014
}
\examples{

}
