% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-wrap-up-functions.R
\name{trainModelRandomForest}
\alias{trainModelRandomForest}
\title{Uplift Random Forest by Guelman et al.}
\usage{
trainModelRandomForest(Y, Trt, X)
}
\arguments{
\item{Y}{is the binary response varibale}

\item{Trt}{is the binary treatment variable}

\item{X}{is the numeric covariates matrix}
}
\value{
a model that is able to classify clients into treatment and control group
}
\description{
Uplift Random Forest by Guelman et al.
}
