% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{trainTianGLM}
\alias{trainTianGLM}
\title{Train function for modified covariate method using generalized linear model}
\usage{
trainTianGLM(Y, Trt, X, opts = NULL)
}
\arguments{
\item{Y}{is the binary response varibale}

\item{Trt}{is the binary treatment variable}

\item{X}{is the numeric covariates matrix}

\item{opts}{is a list of options or list of model parameters such as}
}
\value{
a model that is able to classify clients into treatment and control group
}
\description{
This function fits generalized linear model where covariate transformation as defined by Tian et al., 2014 is used
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fitting model
trainTianGLM(Y, Trt, X, opts = list(alpha = 0, lambda = 0.15))

}
