% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{trainWeisbergRF}
\alias{trainWeisbergRF}
\title{Train Weisberg Random Forest}
\usage{
trainWeisbergRF(Y, Trt, X, opts = NULL)
}
\arguments{
\item{Y}{is the binary response varibale}

\item{Trt}{is the binary treatment variable}

\item{X}{is the numeric covariates matrix}

\item{opts}{is a list of options}
}
\value{
a model that is able to classify clients into treatment and control group
}
\description{
Train Weisberg Random Forest
}
\examples{

}
