% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{trainWeisbergXGb}
\alias{trainWeisbergXGb}
\title{Train function for modified outcome method using XGBoost}
\usage{
trainWeisbergXGb(Y, Trt, X, opts = NULL)
}
\arguments{
\item{Y}{is the binary response varibale}

\item{Trt}{is the binary treatment variable}

\item{X}{is the numeric covariates matrix}

\item{opts}{is a list of options or list of model parameters such as nrounds, eta, subsample, depth}
}
\value{
a model that is able to classify clients into treatment and control group
}
\description{
This function fits XGBoost model where binary response transformation as defined by Weisberg et al., 2015 is used
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fitting model
trainWeisbergXGb(Y, Trt, X, opts = list(nrounds = 15, eta = 0.3, subsample = 0.5, depth = 4))

}
