% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tune.R
\name{trainWlogit_tune}
\alias{trainWlogit_tune}
\title{Tune function for Weisberg GLM}
\usage{
trainWlogit_tune(Y, Trt, X, opts = NULL)
}
\arguments{
\item{Y}{is the binary response varibale}

\item{Trt}{is the binary treatment variable}

\item{X}{is the numeric covariates matrix}

\item{opts}{is a list of options}
}
\value{
a model that is able to classify clients into treatment and control group
}
\description{
Tune function for Weisberg GLM
}
\examples{

}
