% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tune.R
\name{tune2MRF}
\alias{tune2MRF}
\title{Tune function for Random Forest models where difference score method is used (Tune function for *train2MRF*)}
\usage{
tune2MRF(Y, Trt, X, opts = NULL)
}
\arguments{
\item{Y}{the binary response varibale}

\item{Trt}{the binary treatment variable}

\item{X}{the numeric covariates matrix}

\item{opts}{a list of options over that parameters of generalized linear models is chosen (NOTE: should be only ONE common list of parameters for the two models, see Example)}
}
\value{
two lists of parameters found over grid search that then are passed in TrainFunc, one list of parameters for model based on the treated observations (Trt = 1), and one list of parameters for model based on the control observations (Trt = 0)
}
\description{
This function tunes parameters of Random Forest models  using a grid search over supplied parameter ranges such as mtry, ntree, nodesize. Found parameters are passed in TrainFunc
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fiting model
tune2MRF(Y, Trt, X, opts = list(mtry = c(1:2), ntree = c(5, 10, 15), nodesize = c(100, 200, 500)))

}
