% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tune.R
\name{tune2MSVM}
\alias{tune2MSVM}
\title{Tune function for SVM models where difference score method is used (Tune function for *train2MSVM*)}
\usage{
tune2MSVM(Y, Trt, X, opts = NULL)
}
\arguments{
\item{Y}{the binary response varibale}

\item{Trt}{the binary treatment variable}

\item{X}{the numeric covariates matrix}

\item{opts}{a list of options over that parameters of generalized linear models is chosen (NOTE: should be only ONE common list of parameters for the two models, see Example)}
}
\value{
two lists of parameters found over grid search that then are passed in TrainFunc, one list of parameters for model based on the treated observations (Trt = 1), and one list of parameters for model based on the control observations (Trt = 0)
}
\description{
This function tunes parameters of SVM models using a grid search over supplied parameter ranges such as scale, kernel, subset. Found parameters are passed in TrainFunc
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fiting model
tune2MSVM(Y, Trt, X, opts = list(scale = c(F, T), kernel = "linear", subset = c(0.05, 0.15, 0.2)))

}
