% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tune.R
\name{tuneWeisbergGLM}
\alias{tuneWeisbergGLM}
\title{Tune function for generalized linear model where binary response transformation as defined by Weisberg et al., 2015 is used (Tune function for *trainWeisbergGLM*)}
\usage{
tuneWeisbergGLM(Y, Trt, X, opts = NULL)
}
\arguments{
\item{Y}{is the binary response varibale}

\item{Trt}{is the binary treatment variable}

\item{X}{is the numeric covariates matrix}

\item{opts}{is a list of options over that parameters of generalized linear model is chosen}
}
\value{
parameters found over grid search that then are passed in TrainFunc
}
\description{
This function tunes parameters of generalized linear model using a grid search over supplied parameter ranges such as alpha, lambda.
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fiting model
tuneWeisbergGLM(Y, Trt, X, opts = list(alpha = c(0, 1), lambda = c(0.05, 0.15, 0.2)))

}
