% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tune.R
\name{tuneWeisbergRF}
\alias{tuneWeisbergRF}
\title{Tune function for Random Forest model where binary response transformation as defined by Weisberg et al., 2015 is used (Tune function for *trainWeisbergRF*)}
\usage{
tuneWeisbergRF(Y, Trt, X, opts = NULL)
}
\arguments{
\item{Y}{is the binary response varibale}

\item{Trt}{is the binary treatment variable}

\item{X}{is the numeric covariates matrix}

\item{opts}{is a list of options over that parameters of Random Forest model is chosen}
}
\value{
parameters found over grid search that then are passed in TrainFunc
}
\description{
This function tunes parameters of Random Forest model using a grid search over supplied parameter ranges such as mtry, ntree, nodesize.
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fiting model
tuneWeisbergRF(Y, Trt, X, opts = list(mtry = c(1:2), ntree = c(5, 10, 15), nodesize = c(100, 200, 500)))

}
