% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluateAlgos}
\alias{evaluateAlgos}
\title{Evaluates several algos by the given train-test split}
\usage{
evaluateAlgos(models, subgroupQualityFuncs, trainY, trainTrt, trainX, testY,
  testTrt, testX)
}
\arguments{
\item{models}{is a vector of model descriptios that contains Train function, Train options, Predict function and the name of the model.
The prototype of the train function is function(Y,Trt,X, opts),
  where Y is the response variable, Trt is the 0-1 treatment variable, X is the covariate matrix, and opts is the options from the corresponding model.
The prototype for the Predict function is function(m,X) where m is a trained model and X the observations for prediction.}

\item{subgroupQualityFuncs}{is a vector of functions that evaluate the quality of a subgroup. The prototype is function(subgroup, Y, Trt), where subgroup is T-F vector defining a subgroup, and Y and Trt are similar as for the functions from trainModelFuncs}

\item{trainY}{a train response variable}

\item{trainTrt}{a train treatment 0-1 variable}

\item{trainX}{train covariates}

\item{testY}{a test response variable}

\item{testTrt}{a test treatment 0-1 variable}

\item{testX}{test covariates}
}
\value{
a list with found subgroups in the test set, their sizes, qualities, and qualities of a random subset of similar size for all the algorithms
}
\description{
Evaluates several algos by the given train-test split
}
\details{
Takes a set of models and returns the quality of the selected groups by means of subgroupQualityFunc
}
\examples{

# Generating dataset
Trt = rbinom(1000,1,0.5)
X = data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(1000) ) > 0 )
# Defining models
models=list(
  list(Name="RandomForest", TrainFunc=trainModelRandomForest, PredictFunc=predictByModelRandomForest, Opts=NULL),
  list(Name="LMbyTian", TrainFunc=trainModelModLM, PredictFunc=predictByModelModLM, Opts=NULL)
)
# Evaluating algos
res = evaluateAlgos(
    models, # The description of the evaluated models
    c(subgroupAverageTreatmentEffect,subgroupTotalTreatmentEffect), # The set of functions that compute the quality of a subgroup
    Y[1:800], Trt[1:800], X[1:800,], # Train dataset
    Y[801:1000], Trt[801:1000], X[801:1000,] # Holdout dataset
    )
print(res$Qualities)

}
