% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-wrap-up-functions.R
\name{modifyDataByWeisberg}
\alias{modifyDataByWeisberg}
\title{Outcome transformation as defined by Weisberg}
\usage{
modifyDataByWeisberg(Y, Trt)
}
\arguments{
\item{Y}{the binary response varibale}

\item{Trt}{the binary treatment varibale}
}
\value{
the modified outcome vector
}
\description{
Outcome transformation as defined by Weisberg
}
