% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict2MSVM}
\alias{predict2MSVM}
\title{Predict function for train2MSVM}
\usage{
predict2MSVM(m, X)
}
\arguments{
\item{m}{a model trained with train2MSVM}

\item{X}{is the covariate matrix}
}
\value{
a vector with predicted PTE
}
\description{
Predicts PTE of observations by means difference in predictions between two independent SVM models (Predict function for train2MSVM)
}
\examples{

# Generating dataset
N = 1000
Trt = rbinom(N,1,0.5)
X = data.frame(X1=rbinom(N,1,0.6), X2=rnorm(N), X3=rnorm(N))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(N) ) > 0 )
# Fitting model
m = train2MSVM(Y, Trt, X, opts = list(Trt = list(scale = F, kernel = "linear", subset = 0.5), Non_Trt = list(scale = T, kernel = "linear", subset = 0.3)))
# Predicting PTE
predict2MSVM(m, data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000)))

}
