% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-wrap-up-functions.R
\name{predictByModelModLM}
\alias{predictByModelModLM}
\title{Predicts if an observation should be treated or not by means of the model trained with trainModelModLM}
\usage{
predictByModelModLM(m, X)
}
\arguments{
\item{m}{a model trained with trainModelModLM}

\item{X}{is the covariate matrix}
}
\value{
a binary vector specifying the observations to be treated
}
\description{
Predicts if an observation should be treated or not by means of the model trained with trainModelModLM
}
