% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predictUpliftModelRF}
\alias{predictUpliftModelRF}
\title{Predicts if an observation should be treated or not by means of the model}
\usage{
predictUpliftModelRF(m, X)
}
\arguments{
\item{m}{a model trained with trainModelRandomForest}

\item{X}{is the covariate matrix}
}
\value{
a binary vector specifying the observations to be treated
}
\description{
Predicts if an observation should be treated or not by means of the model
}
\examples{

}
