% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-wrap-up-functions.R
\name{toNumericTable}
\alias{toNumericTable}
\title{Transformations of factors to logical variables}
\usage{
toNumericTable(X)
}
\arguments{
\item{X}{is the covariate matrix}
}
\value{
returns a numeric-only covariate matrix
}
\description{
Transformations of factors to logical variables
}
