% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-wrap-up-functions.R
\name{trainModelRandomForest}
\alias{trainModelRandomForest}
\title{Uplift Random Forest by Guelman et al.}
\usage{
trainModelRandomForest(Y, Trt, X, opts)
}
\arguments{
\item{Y}{is the binary response varibale}

\item{Trt}{is the binary treatment variable}

\item{X}{is the numeric covariates matrix}

\item{opts}{is a list of options}
}
\value{
a model that is able to classify clients into treatment and control group
}
\description{
Uplift Random Forest by Guelman et al.
}
\examples{

# Generating dataset
Trt = rbinom(1000,1,0.5)
X = data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000))
Y = as.numeric( ( 2*X$X1 - 1 + X$X2*Trt + rnorm(1000) ) > 0 )
# Training
m = trainModelRandomForest(Y,Trt,X)
# Predicting a good group for new data
sum(predictByModelRandomForest(m, data.frame(X1=rbinom(1000,1,0.6), X2=rnorm(1000), X3=rnorm(1000))))

}
